made changes to the shell.c stopped adding the file name to the PATH some other features were added

+++ _putchar.c
@@ -3,9 +3,10 @@
 /**
  * _putchar - prints out a character
  * @c: input character
+ * * @file_descriptor: carries which output the file should go
  * Return: int on success
  */
 int _putchar(int file_descriptor, char c)
 {
-	return (write(file_descriptor, &c, 1));
+	return (write(file_descriptor, &c, (size_t)1));
 }
